+++ packages/common/src/utils/getToggleElementOnKeyDown.ts
@@ -1,16 +1,17 @@
 import {
   getSlatePluginType,
+  KeyboardHandler,
   mapSlatePluginKeysToOptions,
-  OnKeyDown,
+  SPEditor,
 } from '@udecode/slate-plugins-core';
 import isHotkey from 'is-hotkey';
 import { castArray } from 'lodash';
 import { toggleNodeType } from '../transforms/toggleNodeType';
 import { ELEMENT_DEFAULT } from '../types/node.types';
 
-export const getToggleElementOnKeyDown = (
+export const getToggleElementOnKeyDown = <T extends SPEditor = SPEditor>(
   pluginKeys: string | string[]
-): OnKeyDown => (editor) => (e) => {
+): KeyboardHandler<T> => (editor) => (e) => {
   const defaultType = getSlatePluginType(editor, ELEMENT_DEFAULT);
 
   const options = mapSlatePluginKeysToOptions(editor, pluginKeys);
@@ -21,7 +22,7 @@ export const getToggleElementOnKeyDown = (
     const hotkeys = castArray(hotkey);
 
     for (const key of hotkeys) {
-      if (isHotkey(key, e)) {
+      if (isHotkey(key, e as any)) {
         e.preventDefault();
         toggleNodeType(editor, {
           activeType: type,