+++ packages/core/src/hooks/useEditorRef.ts
@@ -1,0 +1,13 @@
+import { HistoryEditor } from 'slate-history/dist/history-editor';
+import { ReactEditor, useSlateStatic } from 'slate-react';
+import { SPEditor } from '../types/SPEditor';
+import { TEditor } from '../types/TEditor';
+
+/**
+ * Typed {@link useSlateStatic} & SPEditor.
+ * Needs to be called in a child component of `SlatePlugins`.
+ * Else, use `useStoreEditorRef`.
+ */
+export const useEditorRef = <
+  T extends TEditor = ReactEditor & HistoryEditor
+>() => (useSlateStatic() as unknown) as T & SPEditor;