+++ packages/core/src/hooks/useSlatePluginOptions.ts
@@ -1,9 +1,0 @@
-import { getSlatePluginOptions } from '../utils/getSlatePluginOptions';
-import { useTSlateStatic } from './useTSlateStatic';
-
-/**
- * @see {@link getSlatePluginOptions}
- */
-export const useSlatePluginOptions = (pluginKey: string) => {
-  return getSlatePluginOptions(useTSlateStatic(), pluginKey);
-};