+++ packages/core/src/store/useSlatePluginsStore.spec.ts
@@ -1,5 +1,0 @@
-import { slatePluginsStore } from './useSlatePluginsStore';
-
-it('should be', () => {
-  expect(slatePluginsStore.getState()).toBeDefined();
-});