+++ packages/core/src/types/SlatePlugin/OnChange.ts
@@ -3,7 +3,7 @@ import { TNode } from '../TNode';
 
 /**
  * Function called whenever a change occurs in the editor.
- * To prevent the next handler from running, return false.
+ * Return `false` to prevent calling the next plugin handler.
  * @see {@link SlatePropsOnChange}
  */
 export type OnChange<T extends SPEditor = SPEditor> = (