+++ packages/elements/list/src/getListInsertBreak.ts
@@ -40,7 +40,7 @@ export const getListInsertBreak = (editor: SPEditor) => {
         onReset: (_editor) => unwrapList(_editor as SPEditor),
       },
     ],
-  })(editor)(null);
+  })(editor)(new KeyboardEvent('keydown') as any);
   if (didReset) return true;
 
   /**